+++ awjf_waldplan_pub/processing_data/awjf_waldfunktion_processing.sql
@@ -45,4 +45,4 @@ LEFT JOIN awjf_waldplan_pub_v2.t_ili2db_dataset AS dataset
 LEFT JOIN awjf_waldplan_pub_v2.t_ili2db_basket AS basket
 	ON dataset.t_id = basket.dataset
 WHERE
-	wf.t_datasetname = ${bfsnr_param}
+	wf.t_datasetname::int4 = ${bfsnr_param}
